requirements.txt
telebot
openai
python-dotenv
langchain